Re-generate market after hyperjump
Original implementation does not regen market when changing galaxy, but this seems like an oversight since the player has changed planet by doing so. Since we derive price info from the current planet info on the fly, and this changes following a hyperjump, we need to make sure the rest of the market is re-generated in sync.